backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
pandas
meteostat
tqdm
scikit-learn
numpy
openrouteservice
pydantic
typing
sqlalchemy
psycopg2-binary  # or psycopg2
passlib[bcrypt]  # for password hashing
PyJWT            # for JWT creation/verification
pydantic[email]
fastapi-mail
fcntl